+++ lib/textus/manifest/entry/base.rb
@@ -44,6 +44,7 @@ module Textus
         def inject_boot    = false # rubocop:disable Naming/PredicateMethod
         def events         = {}
         def publish_each   = nil
+        def publish_tree   = nil
         def index_filename = nil
         def ignore         = []