+++ lib/textus/manifest/entry/validators.rb
@@ -5,6 +5,7 @@ module Textus
         REGISTERED = [
           Events,
           PublishEach,
+          PublishTree,
           InjectBoot,
           IndexFilename,
           Ignore,